requirements.txt
alembic==1.7.7
aniso8601==9.0.1
atomicwrites==1.4.0
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
coverage==6.3.2
distlib==0.3.4
dnspython==2.2.1
email-validator==1.1.3
execnet==1.9.0
filelock==3.6.0
flake8==4.0.1
Flask==2.1.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
packaging==21.3
pipenv==2022.4.8
platformdirs==2.5.1
pluggy==1.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.8
pytest==7.1.1
pytest-cov==3.0.0
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2
SQLAlchemy==1.4.35
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
Werkzeug==2.1.1
WTForms==3.0.1
